Feat: added screens, routes and events page.
