--- src/App.tsx
@@ -1,18 +1,0 @@
-import React from 'react';
-import Header from './components/Header';
-
-//typescript
-//sass
-//React [hooks]
-
-function App() {
-  return (
-    <>
-      <Header />
-    </>
-  );
-}
-
-export default App;
-
-//yarn add react-router-dom
