+++ src/screens/events.tsx
@@ -1,0 +1,13 @@
+import React from 'react'
+import NavBar from '../components/NavBar'
+
+function Events() {
+    return (
+        <>
+            <NavBar />
+            <div>events</div>
+        </>
+    )
+}
+
+export default Events
