+++ src/screens/landing.tsx
@@ -1,0 +1,13 @@
+import React from 'react'
+import NavBar from '../components/NavBar'
+
+function Landing() {
+    return (
+        <>
+            <NavBar />
+            <div>landing</div>
+        </>
+    )
+}
+
+export default Landing
